Deduplicate some CC test fixtures into conftest.
