+++ tests/cc/test_cc_dataset.py
@@ -1,51 +1,16 @@
 import json
 import shutil
-from datetime import date
 from pathlib import Path
 from tempfile import TemporaryDirectory
 
 import pytest
-import tests.data.cc.dataset
 
 from sec_certs import constants
 from sec_certs.dataset import CCDataset
 from sec_certs.sample.cc import CCCertificate
 
 
-@pytest.fixture(scope="module")
-def data_dir() -> Path:
-    return Path(tests.data.cc.dataset.__path__[0])
-
-
-@pytest.fixture(scope="module")
-def crt() -> CCCertificate:
-    return CCCertificate(
-        "active",
-        "Access Control Devices and Systems",
-        "NetIQ Identity Manager 4.7",
-        "NetIQ Corporation",
-        "SE",
-        {"ALC_FLR.2", "EAL3+"},
-        date(2020, 6, 15),
-        date(2025, 6, 15),
-        "https://www.commoncriteriaportal.org/files/epfiles/Certification%20Report%20-%20NetIQ®%20Identity%20Manager%204.7.pdf",
-        "https://www.commoncriteriaportal.org/files/epfiles/ST%20-%20NetIQ%20Identity%20Manager%204.7.pdf",
-        "https://www.commoncriteriaportal.org/files/epfiles/Certifikat%20CCRA%20-%20NetIQ%20Identity%20Manager%204.7_signed.pdf",
-        "https://www.netiq.com/",
-        set(),
-        set(),
-        None,
-        None,
-        None,
-    )
-
-
-@pytest.fixture
-def toy_dataset(data_dir: Path) -> CCDataset:
-    return CCDataset.from_json(data_dir / "toy_dataset.json")
-
-
-def test_download_and_convert_pdfs(toy_dataset: CCDataset, data_dir: Path):
+def test_download_and_convert_pdfs(toy_dataset: CCDataset, dataset_data_dir: Path):
     template_report_pdf_hashes = {
         "309ac2fd7f2dcf17": "774c41fbba980191ca40ae610b2f61484c5997417b3325b6fd68b345173bde52",
         "8cf86948f02f047d": "533a5995ef8b736cc48cfda30e8aafec77d285511471e0e5a9e8007c8750203a",
@@ -84,8 +49,8 @@ def test_download_and_convert_pdfs(toy_dataset: CCDataset, data_dir: Path):
             assert cert.state.report_txt_path.exists()
             assert cert.state.st_txt_path.exists()
 
-        template_report_txt_path = data_dir / "report_309ac2fd7f2dcf17.txt"
-        template_st_txt_path = data_dir / "target_309ac2fd7f2dcf17.txt"
+        template_report_txt_path = dataset_data_dir / "report_309ac2fd7f2dcf17.txt"
+        template_st_txt_path = dataset_data_dir / "target_309ac2fd7f2dcf17.txt"
         assert (
             abs(toy_dataset["309ac2fd7f2dcf17"].state.st_txt_path.stat().st_size - template_st_txt_path.stat().st_size)
             < 1000
@@ -99,13 +64,13 @@ def test_download_and_convert_pdfs(toy_dataset: CCDataset, data_dir: Path):
         )
 
 
-def test_dataset_to_json(toy_dataset: CCDataset, data_dir: Path, tmp_path: Path):
+def test_dataset_to_json(toy_dataset: CCDataset, dataset_data_dir: Path, tmp_path: Path):
     toy_dataset.to_json(tmp_path / "dset.json")
 
     with (tmp_path / "dset.json").open("r") as handle:
         data = json.load(handle)
 
-    with (data_dir / "toy_dataset.json").open("r") as handle:
+    with (dataset_data_dir / "toy_dataset.json").open("r") as handle:
         template_data = json.load(handle)
 
     del data["timestamp"]
@@ -113,8 +78,8 @@ def test_dataset_to_json(toy_dataset: CCDataset, data_dir: Path, tmp_path: Path)
     assert data == template_data
 
 
-def test_dataset_from_json(toy_dataset: CCDataset, data_dir: Path):
-    assert toy_dataset == CCDataset.from_json(data_dir / "toy_dataset.json")
+def test_dataset_from_json(toy_dataset: CCDataset, dataset_data_dir: Path):
+    assert toy_dataset == CCDataset.from_json(dataset_data_dir / "toy_dataset.json")
 
 
 def test_build_empty_dataset():
@@ -128,12 +93,12 @@ def test_build_empty_dataset():
     assert not dset.state.certs_analyzed
 
 
-def test_build_dataset(data_dir: Path, crt: CCCertificate, toy_dataset: CCDataset):
+def test_build_dataset(dataset_data_dir: Path, cert_one: CCCertificate, toy_dataset: CCDataset):
     with TemporaryDirectory() as tmp_dir:
         dataset_path = Path(tmp_dir)
         (dataset_path / "web").mkdir()
-        shutil.copyfile(data_dir / "cc_products_active.csv", dataset_path / "web" / "cc_products_active.csv")
-        shutil.copyfile(data_dir / "cc_products_active.html", dataset_path / "web" / "cc_products_active.html")
+        shutil.copyfile(dataset_data_dir / "cc_products_active.csv", dataset_path / "web" / "cc_products_active.csv")
+        shutil.copyfile(dataset_data_dir / "cc_products_active.html", dataset_path / "web" / "cc_products_active.html")
 
         dset = CCDataset({}, dataset_path, "sample_dataset", "sample dataset description")
         dset.get_certs_from_web(
@@ -142,7 +107,7 @@ def test_build_dataset(data_dir: Path, crt: CCCertificate, toy_dataset: CCDatase
 
         assert len(list(dataset_path.iterdir())) == 0
         assert len(dset) == 3
-        assert crt in dset
+        assert cert_one in dset
         assert dset == toy_dataset
