+++ tests/cc/test_cc_schemes.py
@@ -17,6 +17,7 @@ def absolute_urls(results):
 
 
 @pytest.mark.xfail(reason="May fail due to server errors.", raises=RequestException)
+@pytest.mark.slow
 def test_australia():
     ineval = CCSchemeDataset.get_australia_in_evaluation()
     assert len(ineval) != 0
@@ -24,6 +25,7 @@ def test_australia():
 
 
 @pytest.mark.xfail(reason="May fail due to server errors.", raises=RequestException)
+@pytest.mark.slow
 def test_canada():
     certified = CCSchemeDataset.get_canada_certified()
     assert len(certified) != 0
@@ -50,6 +52,7 @@ def test_bsi():
 
 
 @pytest.mark.xfail(reason="May fail due to server errors.", raises=RequestException)
+@pytest.mark.slow
 def test_india():
     certified = CCSchemeDataset.get_india_certified()
     assert len(certified) != 0
@@ -60,6 +63,7 @@ def test_india():
 
 
 @pytest.mark.xfail(reason="May fail due to server errors.", raises=RequestException)
+@pytest.mark.slow
 def test_italy():
     certified = CCSchemeDataset.get_italy_certified()
     assert len(certified) != 0
@@ -70,6 +74,7 @@ def test_italy():
 
 
 @pytest.mark.xfail(reason="May fail due to server errors.", raises=RequestException)
+@pytest.mark.slow
 def test_japan():
     certified = CCSchemeDataset.get_japan_certified()
     assert len(certified) != 0
@@ -83,6 +88,7 @@ def test_japan():
 
 
 @pytest.mark.xfail(reason="May fail due to server errors.", raises=RequestException)
+@pytest.mark.slow
 def test_malaysia():
     certified = CCSchemeDataset.get_malaysia_certified()
     assert len(certified) != 0
@@ -93,6 +99,7 @@ def test_malaysia():
 
 
 @pytest.mark.xfail(reason="May fail due to server errors.", raises=RequestException)
+@pytest.mark.slow
 def test_netherlands():
     certified = CCSchemeDataset.get_netherlands_certified()
     assert len(certified) != 0
@@ -103,6 +110,7 @@ def test_netherlands():
 
 
 @pytest.mark.xfail(reason="May fail due to server errors.", raises=RequestException)
+@pytest.mark.slow
 def test_norway():
     certified = CCSchemeDataset.get_norway_certified()
     assert len(certified) != 0
@@ -113,6 +121,7 @@ def test_norway():
 
 
 @pytest.mark.xfail(reason="May fail due to server errors.", raises=RequestException)
+@pytest.mark.slow
 def test_korea():
     certified = CCSchemeDataset.get_korea_certified()
     assert len(certified) != 0
@@ -123,6 +132,7 @@ def test_korea():
 
 
 @pytest.mark.xfail(reason="May fail due to server errors.", raises=RequestException)
+@pytest.mark.slow
 def test_singapore():
     certified = CCSchemeDataset.get_singapore_certified()
     assert len(certified) != 0
@@ -136,6 +146,7 @@ def test_singapore():
 
 
 @pytest.mark.xfail(reason="May fail due to server errors.", raises=RequestException)
+@pytest.mark.slow
 def test_spain():
     certified = CCSchemeDataset.get_spain_certified()
     assert len(certified) != 0
@@ -156,6 +167,7 @@ def test_sweden():
 
 
 @pytest.mark.xfail(reason="May fail due to server errors.", raises=RequestException)
+@pytest.mark.slow
 def test_turkey():
     certified = CCSchemeDataset.get_turkey_certified()
     assert len(certified) != 0
@@ -163,6 +175,7 @@ def test_turkey():
 
 
 @pytest.mark.xfail(reason="May fail due to server errors.", raises=RequestException)
+@pytest.mark.slow
 def test_usa():
     certified = CCSchemeDataset.get_usa_certified()
     for r in certified:
